[WORKAROUND] Prevent bind-now crashes
This is supposed to be a temporary fix only.

This fixes a crash on bind-now objects generated with lld. This seems
to be fixed in a proper way upstream, but so far I did not have the time
to identify and backport these fixes.

Either way, as it is only a check to emit a debugging message, comment
it for now without functional harm.

Signed-off-by: Joao Moreira <joao.moreira@intel.com>